starting ansig
