+++ core/fslibs/parsing_routines/__init__.py
@@ -1,4 +1,4 @@
-#from core.fslibs.parsing_routines.ansig import ansig
+from core.fslibs.parsing_routines.ansig import ansig
 #from core.fslibs.parsing_routines.cara_full import cara_full
 #from core.fslibs.parsing_routines.cara_simple import cara_simple
 from core.fslibs.parsing_routines.ccpnmrv2 import parse_ccpnmrv2_peaklist as ccpnmrv2
